+++ plugins/orchestr8/src/index.ts
@@ -255,9 +255,9 @@ class Orchestr8Server {
               .length,
           },
           searchUri:
-            "@orchestr8://match?query=<keywords>&mode=index&maxResults=5",
+            "orchestr8://match?query=<keywords>&mode=index&maxResults=5",
           usage:
-            "Use @orchestr8://match?query=... for resource discovery. Default mode is 'index' for optimal efficiency.",
+            "Use orchestr8://match?query=... for resource discovery. Default mode is 'index' for optimal efficiency.",
         };
 
         return {